Changed structuring of the coding in the attenuator tool. Added two more adapter files and stored them in an adapters folder.
Added a common functions file for functions common among all tools.

Modified the intermods calculator and power divider design tool to bring them to an operational state.

Added links to the two new tools in the main window.

Images need to be created for the power divider tool and the remaining filters.

There is a known issue in the filter commensurate line calculations - do not use these until it is sorted out.

Update README.md

Adding more To Dos...